Refactored pagination code
